feat(oplog): rename undo module to oplog
