+++ crates/but-api/src/commands/mod.rs
@@ -9,6 +9,7 @@ pub mod git;
 pub mod github;
 pub mod modes;
 pub mod open;
+pub mod oplog;
 pub mod projects;
 pub mod remotes;
 pub mod repo;
@@ -16,7 +17,6 @@ pub mod rules;
 pub mod secret;
 pub mod settings;
 pub mod stack;
-pub mod undo;
 pub mod users;
 pub mod virtual_branches;
 pub mod workspace;
