+++ crates/but/src/oplog/mod.rs
@@ -5,7 +5,7 @@ use gitbutler_project::Project;
 pub(crate) fn show_oplog(project: &Project, json: bool, since: Option<&str>) -> anyhow::Result<()> {
     let snapshots = if let Some(since_sha) = since {
         // Get all snapshots first to find the starting point
-        let all_snapshots = but_api::undo::list_snapshots(project.id, 1000, None, None)?; // Get a large number to find the SHA
+        let all_snapshots = but_api::oplog::list_snapshots(project.id, 1000, None, None)?; // Get a large number to find the SHA
         let mut found_index = None;
 
         // Find the snapshot that matches the since SHA (partial match supported)
@@ -30,7 +30,7 @@ pub(crate) fn show_oplog(project: &Project, json: bool, since: Option<&str>) -> 
             }
         }
     } else {
-        but_api::undo::list_snapshots(project.id, 20, None, None)?
+        but_api::oplog::list_snapshots(project.id, 20, None, None)?
     };
 
     if snapshots.is_empty() {
@@ -119,7 +119,7 @@ pub(crate) fn restore_to_oplog(
     oplog_sha: &str,
     force: bool,
 ) -> anyhow::Result<()> {
-    let snapshots = but_api::undo::list_snapshots(project.id, 1000, None, None)?;
+    let snapshots = but_api::oplog::list_snapshots(project.id, 1000, None, None)?;
 
     // Parse the oplog SHA (support partial SHAs)
     let commit_sha_string = if oplog_sha.len() >= 7 {
@@ -183,7 +183,7 @@ pub(crate) fn restore_to_oplog(
     }
 
     // Restore to the target snapshot using the but-api crate
-    but_api::undo::restore_snapshot(project.id, commit_sha_string)?;
+    but_api::oplog::restore_snapshot(project.id, commit_sha_string)?;
 
     println!("\n{} Restore completed successfully!", "✓".green().bold(),);
 
@@ -197,7 +197,7 @@ pub(crate) fn restore_to_oplog(
 
 pub(crate) fn undo_last_operation(project: &Project, _json: bool) -> anyhow::Result<()> {
     // Get the last two snapshots - restore to the second one back
-    let snapshots = but_api::undo::list_snapshots(project.id, 2, None, None)?;
+    let snapshots = but_api::oplog::list_snapshots(project.id, 2, None, None)?;
 
     if snapshots.len() < 2 {
         println!("{}", "No previous operations to undo.".yellow());
@@ -225,7 +225,7 @@ pub(crate) fn undo_last_operation(project: &Project, _json: bool) -> anyhow::Res
     );
 
     // Restore to the previous snapshot using the but_api
-    but_api::undo::restore_snapshot(project.id, target_snapshot.commit_id.to_string())?;
+    but_api::oplog::restore_snapshot(project.id, target_snapshot.commit_id.to_string())?;
 
     let restore_commit_short = format!(
         "{}{}",
